add gambar di produk

+++ app/Http/Controllers/DashboardController.php
@@ -41,7 +41,6 @@ class DashboardController extends Controller
             ]);
 
             return view('admin.admindashboard');
-
         } elseif (Auth::user()->hasRole('superadmin')) {
             $employe = User::whereRoleIs(['employee'])->get();
             return view('SuperAdmin.admindashboard', [
@@ -179,11 +178,13 @@ class DashboardController extends Controller
      */
     public function edit($id)
     {
-        $user = User::find($id);
-        return view('SuperAdmin.crud.edit', [
-            'user' => $user,
-            'roles' => Role::all(),
-        ]);
+        if (Auth()->hasRole('superadmin')) {
+            $user = User::find($id);
+            return view('SuperAdmin.crud.edit', [
+                'user' => $user,
+                'roles' => Role::all(),
+            ]);
+        }
     }
 
     /**
@@ -204,9 +205,11 @@ class DashboardController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy(User $user)
+    public function destroy(User $user, Item $item)
     {
-        User::destroy($user->id);
-        return redirect('dashboard')->with('success', 'User has been deleted');
+        if (Auth()->hasRole('superadmin')) {
+            User::destroy($user->id);
+            return redirect('dashboard')->with('success', 'User has been deleted');
+        }
     }
 }
